requirements.txt
python-dotenv
sqlalchemy>=2.0.0
psycopg2-binary
praw>=7.0.0